feat: first-pass at module, still has a few issues
This is the first-pass at implementation of `gulp-errorable`, however it
still has a few bugs. The most obvious of which is the fact that it
doesn't have proper support for series, at the moment the tasks are
just fired off and then complete whenever they do, essentially
everything is running in parallel, which is not desired.

Signed-off-by: David Piper <david.piper@mailfence.com>